+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -20053,6 +20053,16 @@ $EC_Completers = {
             break
         }
 
+        # Amazon.ElastiCache.TransitEncryptionMode
+        {
+            ($_ -eq "Edit-ECReplicationGroup/TransitEncryptionMode") -Or
+            ($_ -eq "New-ECReplicationGroup/TransitEncryptionMode")
+        }
+        {
+            $v = "preferred","required"
+            break
+        }
+
 
     }
 
@@ -20069,6 +20079,7 @@ $EC_map = @{
     "NetworkType"=@("New-ECCacheCluster","New-ECReplicationGroup")
     "OutpostMode"=@("New-ECCacheCluster")
     "SourceType"=@("Get-ECEvent")
+    "TransitEncryptionMode"=@("Edit-ECReplicationGroup","New-ECReplicationGroup")
 }
 
 _awsArgumentCompleterRegistration $EC_Completers $EC_map