+++ modules/AWSPowerShell/Cmdlets/ElastiCache/Basic/New-ECCacheCluster-Cmdlet.cs
@@ -434,10 +434,7 @@ namespace Amazon.PowerShell.Cmdlets.EC
         #region Parameter TransitEncryptionEnabled
         /// <summary>
         /// <para>
-        /// <para>A flag that enables in-transit encryption when set to true. You cannot modify the
-        /// value of TransitEncryptionEnabled after the cluster is created. To enable in-transit
-        /// encryption on a cluster you must set <code>TransitEncryptionEnabled</code> to true
-        /// when you create a cluster.</para><para> Only available when creating a cache cluster in an Amazon VPC using Memcached version
+        /// <para>A flag that enables in-transit encryption when set to true.</para><para> Only available when creating a cache cluster in an Amazon VPC using Memcached version
         /// 1.6.12 or later.</para>
         /// </para>
         /// </summary>